Ajusta controller para definir o service e adiciona testes para o controller de delete de comentario

+++ backend/src/controllers/comentario/delete.ts
@@ -1,18 +1,16 @@
-import { Controller, errorResponseHandler } from 'controllers'
+import { errorResponseHandler } from 'controllers'
 import { ComentarioService } from 'services/comentario'
 import { PermissionKeys } from 'types/auth/actors'
 import { Request, Response } from 'types/express'
 import { hasNumericId } from 'utils/request'
+import { ComentarioController } from '.'
 
-export class DeleteComentarioController extends Controller {
-  private service: ComentarioService
-
+export class DeleteComentarioController extends ComentarioController {
   constructor(service: ComentarioService) {
     const validations = [hasNumericId]
     const permission: keyof PermissionKeys = 'comentario_delete'
 
-    super({ validations, permission })
-    this.service = service
+    super({ validations, permission, service })
   }
 
   exec = async (request: Request, response: Response) => {
@@ -20,6 +18,8 @@ export class DeleteComentarioController extends Controller {
       this.validateRequest(request)
 
       const { id } = request.params
+
+      await this.checkUserPermissionScope(request.user, Number(id))
 
       const deletedComentario = await this.service.delete(Number(id))
